Finished e3kit usage with exceptions
