+++ app/src/main/java/com/virgilsecurity/android/e3kitusage/MainActivity.kt
@@ -33,15 +33,12 @@
 
 package com.virgilsecurity.android.e3kitusage
 
-import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.virgilsecurity.android.common.callback.OnGetTokenCallback
 import com.virgilsecurity.android.common.exception.*
 import com.virgilsecurity.android.ethree.interaction.EThree
 import com.virgilsecurity.common.callback.OnCompleteListener
 import com.virgilsecurity.common.model.Data
-import com.virgilsecurity.keyknox.exception.EntryAlreadyExistsException
-import com.virgilsecurity.keyknox.exception.EntryNotFoundException
 import com.virgilsecurity.sdk.common.TimeSpan
 import com.virgilsecurity.sdk.crypto.VirgilAccessTokenSigner
 import com.virgilsecurity.sdk.crypto.VirgilCrypto
@@ -53,54 +50,9 @@ import kotlin.random.Random
 
 class MainActivity : AppCompatActivity() {
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        val identity = UUID.randomUUID().toString()
-        val virgilCrypto = VirgilCrypto()
-
-        val privateKeyData = Data.fromBase64String("MC4CAQAwBQYDK2VwBCIEIPupM43Dt7" +
-                "gJwayKl6EO4qFJbvyALQxap1LcgqoYVREb")
-        val apiKey = virgilCrypto.importPrivateKey(privateKeyData.data).privateKey
-
-        val jwtGenerator = JwtGenerator(
-            "54e071c5c1894aa889e31d6c7864fed5",
-            apiKey,
-            "MCowBQYDK2VwAyEAeAkxVayBD3F4kGQoa1Mtlgqip5jxBXmPG5JP8PXopQI=",
-            TimeSpan.fromTime(600, TimeUnit.SECONDS),
-            VirgilAccessTokenSigner(virgilCrypto)
-        )
-
-        val onGetTokenCallback = object : OnGetTokenCallback {
-            override fun onGetToken(): String {
-                return jwtGenerator.generateToken(identity).stringRepresentation()
-            }
-        }
-
-        val ethree = EThree(identity, onGetTokenCallback, this)
-
-        if (!ethree.hasLocalPrivateKey()) {
-            try {
-                ethree.register().execute()
-
-                ethree.backupPrivateKey(PASSWORD)
-            } catch (exception: AlreadyRegisteredException) {
-                try {
-                    ethree.restorePrivateKey(PASSWORD)
-                } catch (exception: Exception) {
-                    when (exception) {
-                        is EntryNotFoundException, is WrongPasswordException -> {
-                            ethree.rotatePrivateKey()
-
-                            ethree.backupPrivateKey(PASSWORD)
-                        }
-                    }
-                }
-            }
-        }
-        ethree.encrypt(TEXT)
-    }
-
+    // --------------------------------------------------------------------------------------------
+    // variables just to avoid compiler errors ----------------------------------------------------
+    // --------------------------------------------------------------------------------------------
     val rotationConfirmed = Random.nextBoolean()
     val backupPrivateKeyConfirmed = Random.nextBoolean()
     val userSignedIn = Random.nextBoolean()
@@ -127,6 +79,10 @@ class MainActivity : AppCompatActivity() {
 
     lateinit var ethree: EThree
 
+    // -------------------------------------------------------------------------------------------
+    // Beginning of usage example with exceptions description ------------------------------------
+    // -------------------------------------------------------------------------------------------
+
     // This function should fetch Virgil JWT from your backend and return it as Base64 String.
     // You have to provide identity in your request that is required by backend
     // in JwtGenerator#generateToken function.
@@ -137,28 +93,20 @@ class MainActivity : AppCompatActivity() {
         return tokenBase64String
     }
 
-    // Any custom or third-party auth system. (e.g. Firebase)
-    fun signIn() {
-
-    }
-
-    // Any custom or third-party auth system. (e.g. Firebase)
-    fun signUp() {
-
-    }
-
-    // Basic flow contains 3 usual cases:
-    // 1) User already Signed In;
-    // 2) User wants to Sign In;
-    // 3) User wants to Sign Up;
-    //
-    // It's up to developer to determine what's the current case.
-    //
-    // For example you can use Firebase:
-    // call FirebaseAuth#currentUser to check whether the user is Signed In. If user is not
-    // Signed In - call one of:
-    // FirebaseAuth#createUserWithEmailAndPassword or FirebaseAuth#signInWithEmailAndPassword.
-
+    /*
+     * Basic flow contains 3 usual cases:
+     * 1) User is already Signed In;
+     * 2) User wants to Sign In;
+     * 3) User wants to Sign Up;
+     *
+     * It's up to developer to determine what's the current case.
+     *
+     * For example you can use Firebase:
+     * call FirebaseAuth#currentUser to check whether the user is Signed In. If user is not
+     * Signed In - call one of:
+     * FirebaseAuth#createUserWithEmailAndPassword or FirebaseAuth#signInWithEmailAndPassword,
+     * then if operation has finished successful - choose flowTwo or flowThree correspondingly.
+     */
     fun e3kitInitPoint() {
         // The place where all utilities are initialized. (e.g. in Application, MainActivity)
         val onGetTokenCallback = object : OnGetTokenCallback {
@@ -186,77 +134,43 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    // 1) User already Signed In
+    // 1) User is already Signed In.
     fun flowOne() {
-        // Check if the user has private key on device.
+        // Check whether private key is present on the device locally.
         if (ethree.hasLocalPrivateKey()) {
             // If private key of current e3kit user is present - it means all is set up.
             val encrypted = ethree.encrypt("Text") // Encrypted for the current e3kit user herself.
         } else {
-            // There's no private key on device.
-            // There're three common cases:
-            // 1.a) private key is in the Virgil cloud (backed up earlier)
-            // 1.b) private key is on the other device but not in the Virgil cloud (needs to be backed up)
-            // 1.c) private key is lost (there's no backup)
-
-            // Try to restore private key from it's backup in Virgil cloud.
-            ethree.restorePrivateKey(password).addCallback(object : OnCompleteListener {
-                override fun onSuccess() {
-                    // 1.a) private key is in the Virgil cloud (backed up earlier)
-                    // private key has been restored - all is set up.
-                    val encrypted = ethree.encrypt("Text") // Encrypted for the current e3kit user herself.
-                }
-
-                override fun onError(throwable: Throwable) {
-                    if (throwable is PrivateKeyNotFoundException) {
-                        // Actually in this situation you can't know whether it's 1.b or 1.c
-                        // We recommend to ask the user for confirmation that the key has been lost.
-
-                        // 1.b) private key is on the other device but not in the Virgil cloud (needs to be backed up)
-                        // In case user responds that the key has not been lost - ask her to make a
-                        // backup from the device that has private key locally.
-                        // Then try to complete the flowOne again.
-
-                        // In case user responds that the key has been lost - ask user to confirm
-                        // the rotation of keys, so all previous history will become Undecryptable.
-                        // If user confirms that - proceed with 1.c.
-
-                        // 1.c) private key is lost (there's no backup)
-                        if (rotationConfirmed) {
-                            ethree.rotatePrivateKey().addCallback(object : OnCompleteListener {
-                                override fun onSuccess() {
-                                    val encrypted = ethree.encrypt("Text") // Encrypted for the current e3kit user herself.
-                                }
-
-                                override fun onError(throwable: Throwable) {
-                                    // TODO Implement body or it will be empty ):
-                                }
-                            })
-                        }
-                    } else if (throwable is WrongPasswordException) {
-                        // User provided wrong password
-                        // Ask other password and try again flowOne.
-                    } else if (throwable is RestoreKeyException) {
-                        // restorePrivateKey function can restore private key only if there's no one present
-                        // on the device locally, but the private key is already present on the device.
-                        // you have to call ethree.cleanup() first in order to remove local private key.
-                        //
-                        // It's impossible to get here in current flow, but restorePrivateKey function
-                        // can be called in other flows when private key is present on device locally.
-                    }
-                }
-            })
+            tryToRestorePrivateKey()
         }
     }
 
     // 2) User wants to Sign In
     fun flowTwo() {
+        // Check whether private key is present on the device locally.
+        if (ethree.hasLocalPrivateKey()) {
+            // Not recommended.
+            // It's a bad case because you should remove private key from device after session ends.
+            // But generally - it's a possible case, so you can work with e3kit now.
+            val encrypted = ethree.encrypt("Text") // Encrypted for the current e3kit user herself.
+        } else {
+            // Usually on this step you should have your private key removed, but keep the backup in
+            // the Virgil cloud.
+
+            tryToRestorePrivateKey()
+        }
+    }
+
+    // 3) User wants to Sign Up
+    fun flowThree() {
+        // Try to register e3kit user (publishes public key, saves private key on the device locally)
         ethree.register().addCallback(object : OnCompleteListener {
             override fun onSuccess() {
-                // User has been registered successfully. Private key is now on device locally.
+                // User has been registered successfully. Private key is now on the device locally.
                 val encrypted = ethree.encrypt("Text") // Encrypted for the current e3kit user herself.
 
-                // You have to decide to ask the user or to backup private key automatically.
+                // You have to decide to ask the user or to backup the private key automatically.
+
                 if (backupPrivateKeyConfirmed) { // If backup is in user-confirmation flow.
                     ethree.backupPrivateKey(password).addCallback(object : OnCompleteListener {
                         override fun onSuccess() {
@@ -266,12 +180,14 @@ class MainActivity : AppCompatActivity() {
 
                         override fun onError(throwable: Throwable) {
                             if (throwable is BackupKeyException) {
-                                // You get here if there's already private key backup in Virgil cloud
-                                // with current user's identity.
+                                // You get here if there's already private key backup present in the
+                                // Virgil cloud with current user's identity.
                             } else if (throwable is PrivateKeyNotFoundException) {
                                 // You get here if there no private key has been found locally.
+                                //
                                 // It's impossible to get here in current flow, but if you call
-                                // EThree#cleanup before backup - you can get in this situation.
+                                // EThree#cleanup before EThree#backupPrivateKey - you can get into
+                                // this situation.
                             }
                         }
                     })
@@ -280,24 +196,120 @@ class MainActivity : AppCompatActivity() {
 
             override fun onError(throwable: Throwable) {
                 if (throwable is AlreadyRegisteredException) {
-                    // e3kit user with provided identity has been already registered
+                    // e3kit user with provided identity has been already registered.
+                    //
+                    // Possibly you get here because messed up SignIn and SignUp. SignUp should not
+                    // be called if user has been already registered, as well as EThree#register.
                 } else if (throwable is PrivateKeyPresentException) {
-
+                    // This exception means that the private key is already present on the device
+                    // locally.
+                    //
+                    // It's impossible to get this exception while using e3kit only. But you could
+                    // get this exception if you've saved the private key manually before calling
+                    // EThree#register.
+                    // (e.g. using Virgil Core SDK)
                 } else if (throwable is CryptoException) {
-
+                    // This exception could be thrown if there was an exception while generating
+                    // keys during the EThree#register.
+                    //
+                    // The only thing you can do - try to call EThree#register once more.
+                    //
+                    // All known e3kit use-cases do not throw this exception. It's better to contact
+                    // developer in case of this exception.
                 }
             }
         })
     }
 
-    // 3) User wants to Sign Up
-    fun flowThree() {
+    private fun tryToRestorePrivateKey() {
+        // There's no private key on device.
+        //
+        // There're three common cases:
+        // a) private key is in the Virgil cloud (backed up earlier).
+        // b) private key is on the other device but not in the Virgil cloud (needs to be backed up).
+        // c) private key has been lost (there's no backup).
 
-    }
+        // Try to restore private key from it's backup in Virgil cloud.
+        ethree.restorePrivateKey(password).addCallback(object : OnCompleteListener {
+            override fun onSuccess() {
+                // a) private key is in the Virgil cloud (backed up earlier)
+                // private key has been restored - all is set up.
+                val encrypted = ethree.encrypt("Text") // Encrypted for the current e3kit user herself.
+            }
 
+            override fun onError(throwable: Throwable) {
+                if (throwable is NoPrivateKeyBackupException) {
+                    // Actually in this situation you can't know whether it's b) or c).
+                    // We recommend to ask the user for confirmation that the key has been lost.
 
-    companion object {
-        private const val TEXT = "TEXT"
-        private const val PASSWORD = "PASSWORD"
+                    // b) private key is on the other device but not in the Virgil cloud (needs to be backed up).
+                    // In case user responds that the key has Not been lost - ask her to make a
+                    // backup from the device that has private key locally.
+                    // Then try to complete the tryToRestorePrivateKey again.
+
+                    // In case user responds that the key has been lost - ask user to confirm
+                    // the rotation of keys, so all previous history will become Undecryptable.
+                    // If user confirms that - proceed with c).
+
+                    // c) private key has been lost (there's no backup).
+                    if (rotationConfirmed) {
+                        ethree.rotatePrivateKey().addCallback(object : OnCompleteListener {
+                            override fun onSuccess() {
+                                // New public key has been published to the Virgil Cards service.
+                                // New private key has been saved locally on the device.
+                                //
+                                // Don't forget to re-fetch current user's public key for other users
+                                // as previous cached public key has become outdated (if other users have
+                                // this public key cached already).
+                                val encrypted = ethree.encrypt("Text") // Encrypted for the current e3kit user herself.
+                            }
+
+                            override fun onError(throwable: Throwable) {
+                                if (throwable is PrivateKeyPresentException) {
+                                    // rotatePrivateKey function can rotate keys only if there's no private key present
+                                    // on the device locally, but the private key is already present on the device.
+                                    //
+                                    // It's impossible to get here in current flow, but EThree#rotatePrivateKey function
+                                    // can be called in other flows when private key is present on the device locally.
+                                    //
+                                    // You have to call EThree#cleanup in order to remove local private key.
+                                    // As well you can check EThree#hasLocalPrivateKey before calling EThree#rotatePrivateKey.
+                                } else if (throwable is UserNotRegisteredException) {
+                                    // Thrown if EThree#rotatePrivateKey is called before EThree#register.
+                                    // Means that there is no e3kit User registered yet (No public key
+                                    // for current identity in Virgil Cards service is present).
+                                    //
+                                    // You cannot get this exception in current flow. But you can
+                                    // get this exception if you call EThree#rotatePrivateKey before
+                                    // EThree#register has finished successfully.
+                                } else if (throwable is CryptoException) {
+                                    // This exception could be thrown if there was an exception while generating
+                                    // keys during the EThree#register.
+                                    //
+                                    // The only thing you can do - try to call EThree#register once more.
+                                    //
+                                    // All known e3kit use-cases do not throw this exception. It's better to contact
+                                    // developer in case of this exception.
+                                }
+                            }
+                        })
+                    }
+                } else if (throwable is WrongPasswordException) {
+                    // User provided wrong password. (Different from the one that had been used for
+                    // the EThree#backupPrivateKey)
+                    //
+                    // Ask other password and try again to restore private key.
+                } else if (throwable is PrivateKeyPresentException) {
+                    // restorePrivateKey function can restore private key only if there's no one present
+                    // on the device locally, but the private key is already present on the device.
+                    //
+                    // It's impossible to get here in current flow, but EThree#restorePrivateKey function
+                    // can be called in other flows when private key is present on device locally.
+                    //
+                    // You have to call EThree#cleanup in order to remove local private key.
+                    // As well you can check EThree#hasLocalPrivateKey before calling EThree#restorePrivateKey.
+                }
+            }
+        })
     }
 }
